Changing the sleeping pattern of the CSV generator.
My initial impression was that the generator was really sluggish, and it
made me worry that there was something slow about Quix itself. :-o

By changing the way we do sleeping and logging, I think it gives the
same overall positive impression, but also feels like the data is
rocketing through the pipeline.